Finished base for arSequence
All buttons work. There are implemented fail states if the student applies the wrong property at the wrong stage of the plant growth. There is a set up to make the student go through the process as many times as we require. Once required completions have been made, a prompt shows up to move the student to the next scene for the quiz.